fix: improve UX of copy or move function

+++ src-tauri/src/instance/commands.rs
@@ -1,6 +1,6 @@
 use super::{
   super::utils::{
-    fs::copy_whole_dir,
+    fs::{copy_whole_dir, generate_unique_filename},
     nbt::load_nbt,
     path::{get_files_with_regex, get_subdirectories},
   },
@@ -73,7 +73,11 @@ pub fn copy_across_instances(
         None => return Err(InstanceError::InstanceNotFoundByID.into()),
       };
 
-      let dest_path = Path::new(&tgt_path).join(&filename);
+      if !tgt_path.exists() {
+        fs::create_dir_all(&tgt_path).map_err(|_| InstanceError::FolderCreationFailed)?;
+      }
+
+      let dest_path = generate_unique_filename(&tgt_path, &filename);
       fs::copy(&src_file_path, &dest_path).map_err(|_| InstanceError::FileCopyFailed)?;
     }
   } else if src_path.is_dir() {
@@ -83,7 +87,11 @@ pub fn copy_across_instances(
         None => return Err(InstanceError::InstanceNotFoundByID.into()),
       };
 
-      let dest_path = Path::new(&tgt_path).join(src_path.file_name().unwrap());
+      if !tgt_path.exists() {
+        fs::create_dir_all(&tgt_path).map_err(|_| InstanceError::FolderCreationFailed)?;
+      }
+
+      let dest_path = generate_unique_filename(&tgt_path, src_path.file_name().unwrap());
       copy_whole_dir(src_path, &dest_path).map_err(|_| InstanceError::FileCopyFailed)?;
     }
   } else {
@@ -105,12 +113,21 @@ pub fn move_across_instances(
   };
 
   let src_path = Path::new(&src_file_path);
+
+  if !src_path.is_dir() && !src_path.is_file() {
+    return Err(InstanceError::InvalidSourcePath.into());
+  }
+
   let filename = match src_path.file_name() {
     Some(name) => name.to_os_string(),
     None => return Err(InstanceError::InvalidSourcePath.into()),
   };
 
-  let dest_path = Path::new(&tgt_path).join(&filename);
+  if !tgt_path.exists() {
+    fs::create_dir_all(&tgt_path).map_err(|_| InstanceError::FolderCreationFailed)?;
+  }
+
+  let dest_path = generate_unique_filename(&tgt_path, &filename);
   fs::rename(&src_file_path, &dest_path).map_err(|_| InstanceError::FileMoveFailed)?;
   Ok(())
 }
